+++ src/test/resources/db/formula/formula_link_create_mysql.sql
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS formula_links
 (
     formula_link_id bigint       NOT NULL COMMENT 'the internal unique primary index',
     user_id         bigint   DEFAULT NULL COMMENT 'the owner / creator of the formula_link',
-    link_type_id    bigint   DEFAULT NULL,
+    link_type_id    smallint DEFAULT NULL,
     order_nbr       bigint   DEFAULT NULL,
     formula_id      bigint       NOT NULL,
     phrase_id       bigint       NOT NULL,
@@ -32,13 +32,13 @@ ALTER TABLE formula_links
 
 CREATE TABLE IF NOT EXISTS user_formula_links
 (
-    formula_link_id bigint NOT NULL COMMENT 'with the user_id the internal unique primary index',
-    user_id bigint NOT NULL COMMENT 'the changer of the formula_link',
-    link_type_id bigint DEFAULT NULL,
-    order_nbr bigint DEFAULT NULL,
-    excluded smallint DEFAULT NULL COMMENT 'true if a user,but not all,have removed it',
-    share_type_id smallint DEFAULT NULL COMMENT 'to restrict the access',
-    protect_id smallint DEFAULT NULL COMMENT 'to protect against unwanted changes'
+    formula_link_id bigint       NOT NULL COMMENT 'with the user_id the internal unique primary index',
+    user_id         bigint       NOT NULL COMMENT 'the changer of the formula_link',
+    link_type_id    smallint DEFAULT NULL,
+    order_nbr       bigint   DEFAULT NULL,
+    excluded        smallint DEFAULT NULL COMMENT 'true if a user,but not all,have removed it',
+    share_type_id   smallint DEFAULT NULL COMMENT 'to restrict the access',
+    protect_id      smallint DEFAULT NULL COMMENT 'to protect against unwanted changes'
 )
     ENGINE = InnoDB
     DEFAULT CHARSET = utf8
