+++ src/test/resources/db/language/language_create_mysql.sql
@@ -6,7 +6,7 @@
 
 CREATE TABLE IF NOT EXISTS languages
 (
-    language_id    bigint           NOT NULL COMMENT 'the internal unique primary index',
+    language_id    smallint         NOT NULL COMMENT 'the internal unique primary index',
     language_name  varchar(255)     NOT NULL,
     code_id        varchar(100) DEFAULT NULL,
     description    text         DEFAULT NULL,
