+++ src/test/resources/db/verb/verb_create_mysql.sql
@@ -6,7 +6,7 @@
 
 CREATE TABLE IF NOT EXISTS verbs
 (
-    verb_id             bigint           NOT NULL COMMENT 'the internal unique primary index',
+    verb_id             smallint         NOT NULL COMMENT 'the internal unique primary index',
     verb_name           varchar(255)     NOT NULL COMMENT 'the unique type name as shown to the user and used for the selection',
     code_id             varchar(255) DEFAULT NULL COMMENT 'id text to link coded functionality to a specific verb',
     description         text         DEFAULT NULL COMMENT 'text to explain the type to the user as a tooltip; to be replaced by a language form entry',
